Cooking and Cookbook

See merge request 2023ws-se-pr-group/23ws-se-pr-qse-08!20

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/UnitEndpoint.java
@@ -43,13 +43,13 @@ public class UnitEndpoint {
 
     @PermitAll
     @GetMapping("/convert")
-    public Long convertUnits(UnitConvertDto unitConvertDto) throws ValidationException, ConflictException {
+    public Double convertUnits(UnitConvertDto unitConvertDto) throws ValidationException, ConflictException {
         LOGGER.info("convertUnits()");
 
         return unitService.convertUnits(
             unitMapper.unitDtoToEntity(unitConvertDto.from()),
             unitMapper.unitDtoToEntity(unitConvertDto.to()),
-            unitConvertDto.value()
+            unitConvertDto.value().doubleValue()
         );
     }
