+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/ItemListDto.java
@@ -8,9 +8,9 @@ public record ItemListDto(
     @NotEmpty
     String generalName,
     @NotEmpty
-    Long quantityCurrent,
+    Double quantityCurrent,
     @NotEmpty
-    Long quantityTotal,
+    Double quantityTotal,
     @NotEmpty
     Long storId,
     UnitDto unit
